Windows: Fixes subpath volume persistance on container restart
Pods can have volume subpaths mounted inside the containers. Those subpaths can be
based on expressions based on expanded environment variables, which can also be based
on Pod annotations.

Kubernetes expects that the mounted volume subpaths to remain consistent across container
restarts even if the Pod annotation it was based on changed. This behaviour is checked by the
E2E test "should not change the subpath mount on a container restart if the environment variable change"

This test fails on Windows because the subpath itself is always mounted (which is based on the Pod
Annotation), while on Linux, on the first first, the volume subpath is created and mounted under
kubelet/pods/<uid>/volume-subpaths/<volume>/<container-name> and cached, then on restart, it is going
to use the same mount path that already exists, keeping the same subpath mounted in the container.

This commit addresses the issue on Windows, following a similar approach to the Linux implementation.

Refactors some of the code, which can be used on both Linux and Windows.
